move Travis module to its own file
